fix: add suspense to brand page
